Adds support for initialization of unions
Apart from a couple of obvious bug fixes, this focuses on initialization
for unions. The previous initialization code had been designed with only struct
and arrays in mind, so some pretty deep changes had to be made.

   * the data type `Initializer` now has three constructors for regular scalar
     initializers, aggregate initializers (arrays and structs), and union
     initializers. Might want to think about naming `Scalar` differently (see
     section 6.2.5 point 21).
   * `Initializer` is no longer a `Monoid`, since we can't make a `mempty`
     without knowing the type we are initializing. Instead, it is a `Semigroup`,
     with some combinations not valid. This means that either we will need
     to enforce `base >= 4.9` or `semigroups`.

To the best of my current knowledge and testing, this accurately and completely
implements initialization of unions as per C99.